rewrite directives
