+++ src/app/shared/components/bonus-list-container/bonus-list-view/bonus-list-view.component.html
@@ -1,5 +1,9 @@
 <div appTransitionGroup="'flip-list'">
-  <div *ngFor="let bonus of bonuses; trackBy: trackById" class="bonus-container" appTransitionGroupItem>
+  <div
+    *ngFor="let bonus of bonuses; trackBy: trackById"
+    class="bonus-container"
+    appTransitionGroupItem
+  >
     <app-bonus
       [bonus]="bonus"
       [bonusButtonLabel]="bonusButtonLabel"
