+++ src/app/shared/components/bonus-list-container/bonus-list-view/bonus-list-view.component.ts
@@ -13,7 +13,7 @@ export class BonusListViewComponent {
   @Input() bonusButtonLabel: string;
 
   @Output() bonusButtonClickedEvent = new EventEmitter<BonusComponent>();
- 
+
   constructor() {}
 
   public trackById(index: string, item: IBonus): string {
