+++ src/app/shared/directives/transition-group-item.directive.ts
@@ -1,0 +1,16 @@
+import { Directive, ElementRef } from '@angular/core';
+
+@Directive({
+  selector: '[appTransitionGroupItem]',
+})
+export class TransitionGroupItemDirective {
+  prevPos: any;
+  newPos: any;
+  el: HTMLElement;
+  moved: boolean;
+  moveCallback: any;
+
+  constructor(elementRef: ElementRef) {
+    this.el = elementRef.nativeElement;
+  }
+}
