+++ src/app/shared/directives/transition-group.directive.ts
@@ -3,33 +3,16 @@ import {
   AfterContentInit,
   ContentChildren,
   Directive,
-  ElementRef,
   Input,
   QueryList,
-  OnDestroy
+  OnDestroy,
 } from '@angular/core';
+import { TransitionGroupItemDirective } from './transition-group-item.directive';
 
 @Directive({
-  selector: '[appTransitionGroupItem]'
+  selector: '[appTransitionGroup]',
 })
-export class TransitionGroupItemDirective {
-  prevPos: any;
-  newPos: any;
-  el: HTMLElement;
-  moved: boolean;
-  moveCallback: any;
-
-  constructor(elementRef: ElementRef) {
-    this.el = elementRef.nativeElement;
-  }
-}
-
-@Directive({
-  selector: '[appTransitionGroup]'
-})
-
 export class TransitionGroupDirective implements AfterContentInit, OnDestroy {
-
   private changesSubscription: Subscription;
   @Input('appTransitionGroup') class: string;
 
@@ -37,10 +20,9 @@ export class TransitionGroupDirective implements AfterContentInit, OnDestroy {
 
   ngAfterContentInit() {
     this.refreshPosition('prevPos');
-    this.changesSubscription = this.items.changes.subscribe(items => {
+    this.changesSubscription = this.items.changes.subscribe((items) => {
       items.forEach((item: any) => {
         // console.log('item11',item);
-        // console.log('item.newPos11', item.newPos);
         item.prevPos = item.prevPos || item.newPos;
         // console.log('item.prevPos11',item.prevPos)
       });
@@ -49,11 +31,11 @@ export class TransitionGroupDirective implements AfterContentInit, OnDestroy {
       this.refreshPosition('newPos');
       items.forEach(this.applyTranslation);
       this.items.forEach(this.runTransition.bind(this));
-    })
+    });
   }
 
   runCallback(item: TransitionGroupItemDirective) {
-    if(item.moveCallback) {
+    if (item.moveCallback) {
       item.moveCallback();
     }
   }
@@ -63,36 +45,38 @@ export class TransitionGroupDirective implements AfterContentInit, OnDestroy {
       return;
     }
     const cssClass = this.class + '-move';
-    let el = item.el;
-    let style: any = el.style;
+    const el = item.el;
+    const style: any = el.style;
     el.classList.add(cssClass);
     style.transform = style.WebkitTransform = style.transitionDuration = '';
-    el.addEventListener('transitionend', item.moveCallback = (e: any) => {
-      if (!e || /transform$/.test(e.propertyName)) {
-        el.removeEventListener('transitionend', item.moveCallback);
-        item.moveCallback = null;
-        el.classList.remove(cssClass);
-      }
-    });
+    el.addEventListener(
+      'transitionend',
+      (item.moveCallback = (e: any) => {
+        if (!e || /transform$/.test(e.propertyName)) {
+          el.removeEventListener('transitionend', item.moveCallback);
+          item.moveCallback = null;
+          el.classList.remove(cssClass);
+        }
+      }),
+    );
   }
 
   refreshPosition(prop: string) {
-    this.items.forEach(item => {
+    this.items.forEach((item) => {
       item[prop] = {
         top: item.el.offsetTop,
-        left: item.el.offsetLeft
-      }
+        left: item.el.offsetLeft,
+      };
     });
   }
 
   applyTranslation(item: TransitionGroupItemDirective) {
     item.moved = false;
-    // console.log('item.prevPos22',item.prevPos);
     const dx = item.prevPos.left - item.newPos.left;
     const dy = item.prevPos.top - item.newPos.top;
     if (dx || dy) {
       item.moved = true;
-      let style: any = item.el.style;
+      const style: any = item.el.style;
       style.transform = style.WebkitTransform = 'translate(' + dx + 'px,' + dy + 'px)';
       style.transitionDuration = '0s';
     }
