+++ src/app/shared/shared.module.ts
@@ -19,11 +19,7 @@ import { ResizePickDirective } from './directives/resize-pick.directive';
 import { MatRadioModule } from '@angular/material/radio';
 
 @NgModule({
-  declarations: [
-    CloseMenuDirective,
-    LanguageSwitcherDirective,
-    ResizePickDirective,
-  ],
+  declarations: [CloseMenuDirective, LanguageSwitcherDirective, ResizePickDirective],
   imports: [
     RouterModule,
     CommonModule,
